Auswertung der Testwerte eingefügt

+++ main.py
@@ -3,6 +3,6 @@ import numpy as np
 import time
 
 tic = time.time()
-classify("bananas-1-2d", np.linspace(1, 200, 200), 5)
+classify("bananas-1-2d", np.linspace(1, 200, 200, dtype = int), 5)
 toc = time.time()
 print("%.10f seconds" % (toc - tic))
